JSDoc + new components :pencil:
- added JSDocs with correct types
- split off Canvas component
- added WordDiag component
