+++ src/js/components/chat.js
@@ -2,11 +2,20 @@ let element = document.getElementById("messages");
 let form = document.getElementById("gform");
 let props;
 
-export function setup(command, check) {
+/**
+ * Initialize the chat.
+ * @param {Function} command
+ * @param {Function} check
+ */
+export function init(command, check) {
     form.onsubmit = submit;
-    props =  {command, check};
+    props = {command, check};
 }
 
+/**
+ * Submit a chat message.
+ * @param {InputEvent} e
+ */
 function submit(e) {
     e.preventDefault();
     let guessBox = document.getElementById("guessIn");
@@ -21,6 +30,11 @@ function submit(e) {
     guessBox.value = '';
 }
 
+/**
+ * Add a new message from a specific player.
+ * @param {String} text
+ * @param {Player} [player]
+ */
 export function addMessage(text, player) {
     let node = document.createElement("li");
     node.innerHTML = (player ? ('<span style="color:' + player.color + '">'
@@ -28,4 +42,4 @@ export function addMessage(text, player) {
     element.appendChild(node);
 }
 
-export default {setup, addMessage}
+export default {init, addMessage}
