+++ src/js/view.js
@@ -1,5 +1,3 @@
-import {fadeIn} from './tools'
-
 // ----------------------------
 // Info Section
 // ----------------------------
@@ -12,14 +10,3 @@ export function setInfo(text) {
     document.getElementById('info').innerHTML = text;
 }
 
-export function setChoose(words) {
-    try {
-        document.getElementById('word1').value = words[0];
-        document.getElementById('word2').value = words[1];
-        document.getElementById('word3').value = words[2];
-        fadeIn('worddiag');
-    }
-    catch (e) {
-        console.debug("Setting up word dialog -> " + e);
-    }
-}
